fix responsiveness
